Add dyanmic tick scheduling in kernel
Modified:
measure.c has been modified to throw an error if invalid program
name is given

Scheduler has been modified to include dynamic tick
interval scheduling if DYNAMIC_SCHED is defined while
compilation.

The dyanmic tick scheduler increases the tick interval linearly based on
the number of runnable processes.

The number of runnable processes is counted at every iteration of the
scheduler. The tick interval is set to be (NPROC - n) *1000000 cycles,
where NPROC is the max number of processes supported in the kernel, and
n is the number of runnable processes.

Therefore, the lesser the number of runnable processes, the tick
interval is higher to provide for more uninterrupted running time for
each runnable process.

timer.c has been modifed to include a function that updates the interval
value in timer_scratch variable.